requirements.txt
anthropic==0.18.1
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
werkzeug==3.0.1
pdf2image==1.16.3
pillow==10.2.0
pdfplumber==0.10.3
pyyaml==6.0.1
